app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.8.0
alembic==1.13.2
python-dotenv==1.0.1
httpx 
sqlglot 
pyyaml 
pandas
openpyxl
asyncmy
